Make resmgr not startup till configured
With the current setup method, resmgr starts up even before it is configured by
ansible and repeatedly tries to restart. The restarts happen a lot of times
because of the recent migration to systemd.

This change doesn't enable and start resmgr as part of the rpm setup. But we
will rely on the ansible setup to enable and start the service (like other
services).

As an additional change, the service restart policy has been changed to
on-failure mode only.

Review: https://rbcommons.com/s/platform9/r/10459/