Support for ubuntu18 in hostinstaller and hostagent
Necessary changes to get the hostinstaller and hostagent to work on Ubuntu18

Overview of the changes:
- Host installer now allows Ubuntu 18.04 as a supported OS. Updated tests.
- Dropped everything related to init.d scripts in the rpm and deb setups.
Directly setting up systemd because that is the only supported system right now.
This means we are not going to rely on the methods in pf9-service-functions.sh.
- pf9-service-functions.sh will remain to ensure other services which depend on
it continue to work. A new method to check for Ubuntu18 has been added.
- Hacks to get apt working for Ubuntu16 are still applicable in Ubuntu18. So
reverse the condition check to do it for everything other than Ubuntu14.

Testing done:
http://teamcity.platform9.sys:8111/viewLog.html?tab=buildLog&buildTypeId=Pf9project_IntegrationBuildPipeline_001smokeTests&buildId=954117
but ostackhost service is not coming up on the host.

Issues: CORE-712

Review: https://rbcommons.com/s/platform9/r/10686/