Fix resmgr service start interval
The StartLimitIntervalSec parameter is available only in newer versions of
systemd (230 onwards). The systemd we use on the DU is much older (219) where
this parameter is just StartLimitInterval. This change uses the old parameter
for the service.

Testing: Patched a broken DU and ensured these warnings are gone after the fix

Review: https://rbcommons.com/s/platform9/r/10694/