Kill running sidekick instance on hostagent uninstall
Sidekick continues to run after uninstall of comms to ensure that the backup
troubleshooting channel continues to be available even when comms is down. The
side effect of this is that it continues to run on complete uninstall of all
pf9 packages. The user is responsible for cleaning this process up manually.

In this change, we are relying on the assumption that uninstall of hostagent
means
the user is clearing up all associated pf9 services on the host. In the
uninstall
step of hostagent, kill the sidekick process. Note that hostagent is upgraded
using
update and not the uninstall/reinstall model the apps use.

Also call out explicit dependency on procps for ubuntu (was already done for
CentOS)

Testing: http://teamcity.platform9.sys:8111/viewLog.html?buildId=957552

Issues: CORE-713