Tokendecoder-->Tokenextractor
Tokendecoder-->Tokenextractor

In the past, we had extremely large keystone tokens and there was a need to
compress and base64 encode it when passing it to a service. However, with the
move to fernet tokens, tokens became much smaller in length. This means we don't
need the middleware to decode and extract the token anymore. This change drops
the logic to do that. However, we do need the middleware to still handle
requests that have the token only in the cookie and the middleware should put
the token in the environ in such cases.

Motivation for this change is that in py3, the base64 decode changes have a
change in semantics compared to py2. It can be fixed but given there is no need
to handle compressed tokens anymore, it is preferred to drop that to the
middleware.

Changed the middleware name according to its new functionality

Testing done:
With py3 setup: http://teamcity.platform9.sys:8111/viewLog.html?buildId=980410
(failed but the download of the installer where this middleware is used is
successful
With py2 setup: http://teamcity.platform9.sys:8111/viewLog.html?buildId=980359

There is a subsequent change needed for not passing compressed tokens in
pf9-main. That will come as part of a different review.

Issues: CORE-711

Review: https://rbcommons.com/s/platform9/r/10775/