Fix container resmgr container build with new tokenextractor change
A recent change had repurposed the tokendecoder module to a tokenextractor
module. The original paste conf file relating to this was changed but there is a
different conf file when it comes to the resmgr container which was missed out.
This change adds makes the necessary changes on the container side too.

Testing: Validated the container builds now pass

Issues addressed: CORE-711