Drop metrics when a host is deauthorized as well
When a host is offline and it is deauthorized, it will be removed from the
resmgr DB but will not be added into the unauthorized host list (because it is
offline). This would leave the metrics lying around for the offline host because
the metrics for the host are removed only when it transitions into unauthorized
host and then disappears from resmgr (happy path). This change addresses removal
of the metrics for the offline case as well.

Issues: CORE-809

Review: https://rbcommons.com/s/platform9/r/10989/

Testing done: Added node to a cluster. Stopped hostagent. Ensured node is
reporting offline in metrics. Deauthorized the node. Verified that the node
metrics are cleaned up from resmgr metrics endpoint.