Move roles through state machine when offline
A host that is offline can get stuck during deauth (in the pre-deauth state)
because we don't process the host in the state transition algorithm if the host
is _not_ reported by bbmaster. A host will not be reported by bbmaster if
bbmaster has not received a heartbeat (maybe host is offline) since the last
restart of bbmaster. Not being able to deauth such a host leaves the host in a
state which cannot be cleaned up using APIs.

This changes makes the host go through the state transition algorithm if the
host is not deauthorized and _not_ reported by bbmaster. They get classified as
"absent" hosts in the bbone poller loop. This also relies on the assumption that
the pre-deauth hooks would take into account of checking the host is offline if
it needs to prevent proceeding with the deauthorization of such hosts.

Testing done: Patched a PMK setup and ensured a PMK node gets deauthorized even
when it is offline as per the state mentioned above.

Issues: CORE-334

Review: https://rbcommons.com/s/platform9/r/10999/