Prevent serving resmgr/metrics externally for decco environment
The metrics are meant only for internal consumption and should not be exposed to
the external world from our HTTP REST endpoint.

This is an interim fix to plug the security hole quickly. Longer fix is being
planned out.

Issues: CORE-833

Testing done: Patched a deployment and ensured that metrics don't get exposed
anymore.