Move all service commands to use systemctl
All core services are now systemctl compatible. Move the service call in du_ctl
and the installer to use systemctl. Note that the service call defaults to
systemctl on current systems anyway but there are problems with the current
implementation.

1) Using systemctl status always returns a success code even when the service
may be down. The right call to do in this case is systemctl is-active which will
return an error code when the service is in fact not running.

2) systemctl has this odd behavior of using pager mode by default when there is
an active tty. Certain automation may lead to a tty session. Even though cat may
be the default pager, it very likely could be something else. In at least one
customer setup, we noticed terraform was using more. This lead to systemctl
commands hanging till a user quit the pager even in automation mode. Solution is
to use the --no-pager option with systemctl commands.

Issues: CORE-875

Review: https://rbcommons.com/s/platform9/r/11094/