Ensure the CN passed to vouch is less than 64 chars
Vouch can only handle CN of length lesser than 64 characters. This limit is
enforced by Vault which is used by vouch to sign certs. This change implements a
truncation of the CN at the call site.

This is needed because Azure sets up really long hostname and we append the UUID
to the hostname.

Testing done: Patched such a setup with this change and ensured that certs can
now be requested.

Issues: CORE-887

Review: https://rbcommons.com/s/platform9/r/11123/