Introduce hostagent pre-install script for deb
The hostagent.conf file is being used as a config file in the deb package to
manage how this conf file needs to be setup as part of the certless and
traditional hostagent deployment.

However, the transition between a regular file to a conf file needs more
handling to prevert the debian package management system from complaining about
config file conflicts. More details:
https://platform9.atlassian.net/l/c/Np2Vjwih

In this change, we use the pre-install script to detect if the hostagent conf
file is not a config file already in the debian package system. If it isn't,
then we move it a different location to allow the config file from the new
package to be setup. Note that once the config file is marked as a configuration
file, subsequent upgrades should just work with newer config files.

Issues: CORE-319

Testing done: This has been validated in the following cases manually
1. host going from 4.0 to 4.1 with a cert version change
2. host going from 4.1 to 4.2 with a cert version change
3. host going from 4.0 to 4.2 with a cert version change (in case a host misses
the 4.0 to 4.1 transition because it was offline  or we missed to rectify a failed
upgrade)
Also run smoke and upgrade tests

Review: https://rbcommons.com/s/platform9/r/11134/