Ensure that the installer uses CN that meets DNS specification
The CN used in CSR request to vault is hostname+hostuuid (trimmed to 63 chars).
This can lead to odd patterns of the CN in the CSR. Example: CN ending with a -
which doesn't meet the DNS spec. Such requests tend to be rejected by vault.

Instead of above approach, we will now use part of the hostname (first 54 chars)
+ '-' + first 8 chars of the uuid. We will also run it through a regex
validator. If the validator fails, it will warn a message but will proceed to
send the request anyway.

Issues: CORE-948

Testing done: Validated that the bash script logic works well for different
hostname conditions, as well as invalid case.