Resmgr active flag : Fix upgrade case
Description:
Root cause:
Recently, role creation API was added with a support of passing active
state in the payload. (ref: https://rbcommons.com/s/platform9/r/11539/)
This caused a case where during an upgrade,
a role with an older version is already present. Since the new role via
upgrade is not added via an API, it is not marked as active, which resulted
in upgrade failures.

Fix:
Until the logic to push the largest version of a role is implemented,
all roles created via save_role_in_db will be marked as active, and other existing
versions of the same role are marked as inactive.
This will take care of following scenarios:
- New deployment, where role is read from the local FS
- Upgrades, where new version is also read from local FS
- Roles created via API (passing active flag is now supported)
If other scenarios occur, there is already an API provided to mark a role with
a specific version as an active role.
This covers all the cases.

Modules affected: resmgr
RB : https://rbcommons.com/s/platform9/r/11558/