Support for CentOS 8
This change adds in support for CentOS 8 at the host installer, hostagent and
friends services. Changes needed
- Dropped support for CentOS 7.0-7.2 in the host installer
- Added support for CentOS 8.0-8.3 in the host installer
- chrony is the default NTP service in CentOS 8. Use that when the OS is CentOS
8 (will this conflict with ntpd if already installed?)
- Package jq and use it in the installer. The installer was relying on python
and this is not guaranteed to be present by default on the host.
- Drop the requirement/dependency on python-setuptools in hostagent. Don't see
the need for it.
- Replace references to ntpd with ntp. Kept older ntpd option in the installer
for backward compatibility.

Testing done: Deployed a new DU with this change. Ran the host installer.
Ensured hostagent and comms came up fine.
Also
http://teamcity.platform9.horse/buildConfiguration/Pf9project_IntegrationBuildPipeline_001smokeTests/1431811