Don't log extension failures in the hostagent daemon log (#35)
* CORE-1159

Desc:
A failure in extension execution is being logged in hostagent-daemon.log file. In case of
consistent failure, this will fill-up the log file. This log file is not controlled by the
hostagent's log rotation rules.

Fix:
In case of failure in execution of extension, a failure message would logged in the hostagent.log
file and details of failure would be availbale in the extension details.

* Remove unnecessary argument to check_output

* Incorporating review comments