Don't merge the stdout and stderr when running extensions (#46) (#47)
Extensions may write to stderr when some dependent commands do so. These may be
error, warning or any information of the command run. When an extension writes
to stderr, none of that merged output is json serializable. That implies the
extension data is never returned to the DU and an error is returned even though
the extension may have written the right out to stdout (along with some other
data to stderr).

This change in interface is causing assumptions in dependent services like qbert
to break when it sees an extension return error.

A fix here is to keep stdout and stderr separate. Log stderr in hostagent logs
and discard it for now.

Issues: CORE-1175

Testing: Patched staging-ft hosts and noticed things hold up better