Make a deepcopy of app_info for filtering its data (#52)
The app_info object contains data for all apps associated with a host. The code
also trims the du_config part of the app_info. This will affect future
processing of other apps because the trimming happens even for apps not being
pushed to the host currently.

Given that the du_config part should be skipped only when sending data to the
host, this change makes a deepcopy of the object and filters du_config only for
the new copy which is then sent down to the host. This prevents unwanted deletes
of the main app_info data structure.

Issues: CORE-1179

Testing done: All done live in sbux setup