CORE-1200 (#68)
Improve DB handling while advancing role state of a host in resource manager.

Call session.commit only when the row is updated. During PMO scale we observed
resmgr is spending good amount of time session.commit(). Additionally, this function
is called for all the hosts in every poller loop. This increases the poller loop
time in steady state of execution.