Increase the waiting for hostagent dir timeout in installer (#94)
Currently is a 60 sec timeout. Noticed in some setups, it is taking longer than
that for hostagent to query the apps, cert update, package cleanup. Given the
longer startup time here, increase the timeout to 150 to account for more
processing and potentially slower machines where this is run on.

Issues: CORE-1239