Drop unintended return from process_existing_hosts during hosts processing (#99)
There is a code path in _process_existing_hosts where we return from the method
when the info field is missing from bbmaster host data. This field is in fact
missing when the hosts are in missing status in bbmaster. This return prevents
processing any further “existing” hosts in the poller loop which can prevent the
state machine transition for those machines.

This change addresses this incorrect behavior by eliminating that return path
and processing the host info only if the info field is present. If not present,
we move on to the next step/next host to process.

Testing done: Patched a resmgr exhibiting this problem. Ensured that the poller
loop now completes the necessary processing for the host.

Issues: CORE-1243