Add restart of sidekick when certs are changed on the host (#128)
Previous changes involved only comms restart and sidekick restart was left out.
Both the services rely on the same set of certificates which can change as part
of the cert setup and rotation on the host.
This change tags on the sidekick service restart in addition to the pf9-comms
service restart. Note that if either service restart fails, we will return an
error right now. However, the connectivity check is done only through comms. If
the certificate works for comms, it is very likely will work for sidekick as
well (Also there is no good sidekick API to leverage for such a connectivity
test).

Issues: CORE-1269