Add CA rotation logic for hostagent (#123)
Change introduces the logic to hook into the existing cert update thread in
hostagent to also rotate any CAs, if there is an update needed.

As part of the cert update thread, we first check to see if there is a new CA
advertised by vouch. This is done by comparing the CAs locally held vs the ones
advertised by vouch. If they differ, then we write out those CAs locally.
The rest of the cert rotation process follows as earlier which includes restart
of the comms and sidekick service.

Addresses: CORE-946