Add error handling for hostagent.conf download from DU (#136) (#137)
In KubeDU setups, the hostagent.conf is generated on the DU and downloaded on to
the host as part of the hostagent pre-instal script. However, there is no error
handling here in case the curl/http connection fails.

This change explicitly errors out when the command doesn't return a 200 HTTP
code.

Issues: CORE-1281