Adjust hostagent.conf download code to fail on http error (#138) (#139)
Previous implementation was not taking into account that http output code needs
to be inspected. But this also gets complicated because the output is also
expected from the HTTP request. Instead, use the --fail option which should
return an exit code of 22 from curl for _most_ of the curl requests that dont
return in the 2xx range.

Issue: PMK-5443